started on randomize api fetch
